# Conflicts:
#	Computation/constants.h
#	Computation/kernel.h
#	Computation/metrics.h
#	Rendering/Renderer.h

+++ Computation/constants.h
@@ -1,5 +1,5 @@
 #pragma once
 
-constexpr float KERNEL_ATTRACTOR_MAX_TOL = 0.1f;
+constexpr double KERNEL_ATTRACTOR_MAX_TOL = 0.1;
 constexpr int MAX_PERIOD = 5;
 constexpr unsigned int MAIN_KERNEL_BLOCK_SIZE = 1024;
