+++ Computation/metrics.h
@@ -1,7 +1,6 @@
 #pragma once
-#include "kernel_types.h"
 
-// __device__ __inline__ DIST_F(complex z, dist_t maxIters, complex p, float arg, complex c)
+// __device__ __inline__ dist_t DIST_F(complex z, int maxIters, complex p, real arg, complex c)
 
 #ifdef BUILD_FOR_NVRTC
 #include "kernel_macros.cuh"
@@ -45,7 +44,7 @@ enum DistanceMetric {
     PERIODIC
 };
 
-inline float prepMetricArg(DistanceMetric metric, float arg) {
+inline double prepMetricArg(DistanceMetric metric, double arg) {
     if(metric == FIXEDPOINT_ITERATIONS || metric == JULIA || metric == FIXEDPOINT_EUCLIDEAN || metric == ATTRACTOR || metric == PERIODIC)
         return arg * arg;
     return arg;
